requirements.txt
asyncio==3.4.3
beautifulsoup4==4.13.4
click==8.1.8
fastapi==0.115.12
httpx==0.28.1
langchain==0.3.24
langchain-community==0.3.22
numpy==2.2.5
openai==1.76.0
orjson==3.10.16
pandas==2.2.3
peewee==3.17.9
pydantic==2.11.3
requests==2.32.3
sqlalchemy==2.0.40
structlog==25.2.0
tenacity==9.1.2
tqdm==4.67.1
uvicorn==0.34.2
yfinance==0.2.56
zstandard==0.23.0
python-dotenv==1.0.1
tinydb==4.8.2
openai-agents==0.0.13
langchain-openai==0.3.14
pytest==8.3.5
httpx==0.28.1
